chatAppPython/requirements.txt
qrcode[pil]
pycryptodome
requests==2.23.0